refactor, add overloaded functions to work with MouseInput directly witchout explicit static_cast required

+++ src/ncurses_def.cpp
@@ -331,11 +331,11 @@ input_event input_manager::get_input_event( const keyboard_mode /*preferred_keyb
                 rval.type = input_event_t::mouse;
                 rval.mouse_pos = point( event.x, event.y );
                 if( event.bstate & BUTTON1_CLICKED ) {
-                    rval.add_input( static_cast<int>(MouseInput::LeftButtonReleased) );
+                    rval.add_input( MouseInput::LeftButtonReleased );
                 } else if( event.bstate & BUTTON3_CLICKED ) {
-                    rval.add_input( static_cast<int>(MouseInput::RightButtonReleased) );
+                    rval.add_input( MouseInput::RightButtonReleased );
                 } else if( event.bstate & REPORT_MOUSE_POSITION ) {
-                    rval.add_input( static_cast<int>(MouseInput::Move) );
+                    rval.add_input( MouseInput::Move );
                     if( input_timeout > 0 ) {
                         // Mouse movement seems to clear ncurses timeout
                         set_timeout( input_timeout );
